requirements.txt
et-xmlfile==1.1.0
numpy==1.23.1
openpyxl==3.0.10
pandas==1.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
selenium==3.141.0
six==1.16.0
urllib3==1.26.11
